Move src/MaxText/distillation to src/maxtext/trainers/post_train/distillation
