+++ .coveragerc
@@ -10,7 +10,9 @@ omit =
 [paths]
 source =
     src/MaxText
+    src/maxtext
     */site-packages/MaxText
+    */site-packages/maxtext
 
 [report]
 show_missing = True
