+++ src/maxtext/trainers/post_train/distillation/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2023–2026 Google LLC
+# Copyright 2023-2026 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
